requirements.txt
setuptools
pyyaml
wheel
pyinstaller
boto3
awscli
configparser
